make executor and statedb optional
