+++ state/forkid.go
@@ -1,11 +1,19 @@
 package state
 
+import "github.com/0xPolygonHermez/zkevm-node/log"
+
 // ForkIDInterval is a fork id interval
 type ForkIDInterval struct {
 	FromBatchNumber uint64
 	ToBatchNumber   uint64
 	ForkId          uint64
 	Version         string
+}
+
+// UpdateForkIDIntervals updates the forkID intervals
+func (s *State) UpdateForkIDIntervals(intervals []ForkIDInterval) {
+	log.Infof("Updating forkIDs. Setting %d forkIDs", len(intervals))
+	s.cfg.ForkIDIntervals = intervals
 }
 
 // GetForkIDByBatchNumber returns the fork id for a given batch number
@@ -18,4 +26,9 @@ func GetForkIDByBatchNumber(intervals []ForkIDInterval, batchNumber uint64) uint
 
 	// If not found return the last fork id
 	return intervals[len(intervals)-1].ForkId
+}
+
+// GetForkIdByBatchNumber returns the fork id for the given batch number
+func (s *State) GetForkIdByBatchNumber(batchNumber uint64) uint64 {
+	return GetForkIDByBatchNumber(s.cfg.ForkIDIntervals, batchNumber)
 }
