+++ synchronizer/mock_state.go
@@ -7,6 +7,8 @@ import (
 
 	common "github.com/ethereum/go-ethereum/common"
 
+	metrics "github.com/0xPolygonHermez/zkevm-node/state/metrics"
+
 	mock "github.com/stretchr/testify/mock"
 
 	pb "github.com/0xPolygonHermez/zkevm-node/state/runtime/executor/pb"
@@ -470,15 +472,15 @@ func (_m *stateMock) OpenBatch(ctx context.Context, processingContext state.Proc
 }
 
 // ProcessAndStoreClosedBatch provides a mock function with given fields: ctx, processingCtx, encodedTxs, dbTx, caller
-func (_m *stateMock) ProcessAndStoreClosedBatch(ctx context.Context, processingCtx state.ProcessingContext, encodedTxs []byte, dbTx pgx.Tx, caller state.CallerLabel) (common.Hash, error) {
+func (_m *stateMock) ProcessAndStoreClosedBatch(ctx context.Context, processingCtx state.ProcessingContext, encodedTxs []byte, dbTx pgx.Tx, caller metrics.CallerLabel) (common.Hash, error) {
 	ret := _m.Called(ctx, processingCtx, encodedTxs, dbTx, caller)
 
 	var r0 common.Hash
 	var r1 error
-	if rf, ok := ret.Get(0).(func(context.Context, state.ProcessingContext, []byte, pgx.Tx, state.CallerLabel) (common.Hash, error)); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, state.ProcessingContext, []byte, pgx.Tx, metrics.CallerLabel) (common.Hash, error)); ok {
 		return rf(ctx, processingCtx, encodedTxs, dbTx, caller)
 	}
-	if rf, ok := ret.Get(0).(func(context.Context, state.ProcessingContext, []byte, pgx.Tx, state.CallerLabel) common.Hash); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, state.ProcessingContext, []byte, pgx.Tx, metrics.CallerLabel) common.Hash); ok {
 		r0 = rf(ctx, processingCtx, encodedTxs, dbTx, caller)
 	} else {
 		if ret.Get(0) != nil {
@@ -486,7 +488,7 @@ func (_m *stateMock) ProcessAndStoreClosedBatch(ctx context.Context, processingC
 		}
 	}
 
-	if rf, ok := ret.Get(1).(func(context.Context, state.ProcessingContext, []byte, pgx.Tx, state.CallerLabel) error); ok {
+	if rf, ok := ret.Get(1).(func(context.Context, state.ProcessingContext, []byte, pgx.Tx, metrics.CallerLabel) error); ok {
 		r1 = rf(ctx, processingCtx, encodedTxs, dbTx, caller)
 	} else {
 		r1 = ret.Error(1)
@@ -496,15 +498,15 @@ func (_m *stateMock) ProcessAndStoreClosedBatch(ctx context.Context, processingC
 }
 
 // ProcessSequencerBatch provides a mock function with given fields: ctx, batchNumber, batchL2Data, caller, dbTx
-func (_m *stateMock) ProcessSequencerBatch(ctx context.Context, batchNumber uint64, batchL2Data []byte, caller state.CallerLabel, dbTx pgx.Tx) (*state.ProcessBatchResponse, error) {
+func (_m *stateMock) ProcessSequencerBatch(ctx context.Context, batchNumber uint64, batchL2Data []byte, caller metrics.CallerLabel, dbTx pgx.Tx) (*state.ProcessBatchResponse, error) {
 	ret := _m.Called(ctx, batchNumber, batchL2Data, caller, dbTx)
 
 	var r0 *state.ProcessBatchResponse
 	var r1 error
-	if rf, ok := ret.Get(0).(func(context.Context, uint64, []byte, state.CallerLabel, pgx.Tx) (*state.ProcessBatchResponse, error)); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, uint64, []byte, metrics.CallerLabel, pgx.Tx) (*state.ProcessBatchResponse, error)); ok {
 		return rf(ctx, batchNumber, batchL2Data, caller, dbTx)
 	}
-	if rf, ok := ret.Get(0).(func(context.Context, uint64, []byte, state.CallerLabel, pgx.Tx) *state.ProcessBatchResponse); ok {
+	if rf, ok := ret.Get(0).(func(context.Context, uint64, []byte, metrics.CallerLabel, pgx.Tx) *state.ProcessBatchResponse); ok {
 		r0 = rf(ctx, batchNumber, batchL2Data, caller, dbTx)
 	} else {
 		if ret.Get(0) != nil {
@@ -512,7 +514,7 @@ func (_m *stateMock) ProcessSequencerBatch(ctx context.Context, batchNumber uint
 		}
 	}
 
-	if rf, ok := ret.Get(1).(func(context.Context, uint64, []byte, state.CallerLabel, pgx.Tx) error); ok {
+	if rf, ok := ret.Get(1).(func(context.Context, uint64, []byte, metrics.CallerLabel, pgx.Tx) error); ok {
 		r1 = rf(ctx, batchNumber, batchL2Data, caller, dbTx)
 	} else {
 		r1 = ret.Error(1)
